requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.4
parso==0.7.0
pathspec==0.8.0
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.1
typed-ast==1.4.1
urllib3==1.25.9
waitress==1.4.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
